requirements.txt
asgiref==3.7.2
coverage==7.3.2
dj-database-url==2.1.0
Django==4.2.6
django-filter==23.3
django-heroku==0.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.6.0
